More UI conversion to new framework
