+++ gui/app/styles/view/space.scss
@@ -1,18 +1,10 @@
 .view-space {
-	> .heading {
-		height: 10px;
-
-		> .view-heading {
-			font-size: 1.5rem;
-			font-weight: bold;
-			color: $color-dark;
-		}
-	}
-
 	> .filter-caption {
+		margin: 10px 0;
 		color: $color-gray;
-		font-size: 1.0rem;
-		margin: 0 0 10px 0;
+		font-size: 1.1rem;
+		font-weight: 500;
+		text-align: center;
 	}
 
 	> .documents {
@@ -27,7 +19,6 @@
 			position: relative;
 			margin: 0 0 30px 0;
 			width: 100%;
-			// height: 150px;
 
 			&:hover {
 				> .checkbox {
@@ -55,14 +46,6 @@
 
 				&:hover {
 					color: $color-gray;
-
-					> .title {
-						color: $color-link;
-					}
-
-					> .snippet {
-						color: $color-link;
-					}
 				}
 			}
