+++ gui/app/templates/components/folder/space-heading.hbs
@@ -1,28 +1,20 @@
-<div class="container">
-	<div class="row">
-		<div class="col">
-			<div class="mt-4 mb-5">
-				{{#unless editMode}}
-					<div class="view-space">
-						<div class="heading">
-							<h1 class="view-heading {{if permissions.spaceOwner 'cursor-pointer'}}" onclick={{if permissions.spaceOwner (action 'toggleEdit')}}>
-								{{space.name}}
-							</h1>
-						</div>
-					</div>
-				{{else}}
-					<form {{action "onSave" on="submit"}}>
-						<div class="view-space">
-							<div class="heading">
-								<div class="form-group">
-									{{focus-input id="space-name" type="text" value=spaceName class=(if hasNameError 'form-control is-invalid' 'form-control') placeholder="Space name" autocomplete="off"}}
-									<small class="form-text text-muted">Press Enter to save or Escape to cancel</small>
-								</div>
-							</div>
-						</div>
-					</form>
-				{{/unless}}
-			</div>
+{{#unless editMode}}
+	<div class="view-space">
+		<div class="heading">
+			<h1 class="view-heading {{if permissions.spaceOwner 'cursor-pointer'}}" onclick={{if permissions.spaceOwner (action 'toggleEdit')}}>
+				{{space.name}}
+			</h1>
 		</div>
 	</div>
-</div>
+{{else}}
+	<form {{action "onSave" on="submit"}}>
+		<div class="view-space">
+			<div class="heading">
+				<div class="form-group">
+					{{focus-input id="space-name" type="text" value=spaceName class=(if hasNameError 'form-control is-invalid' 'form-control') placeholder="Space name" autocomplete="off"}}
+					<small class="form-text text-muted">Press Enter to save or Escape to cancel</small>
+				</div>
+			</div>
+		</div>
+	</form>
+{{/unless}}
